add post upload with media. add getters for posts by popularity

+++ src/main/java/com/john_deligiannis/laugh_9/bodies/DeletePostRequest.java
@@ -2,13 +2,13 @@ package com.john_deligiannis.laugh_9.bodies;
 
 public class DeletePostRequest {
 	
-	private Integer postId;
+	private Long postId;
 
-	public Integer getPostId() {
+	public Long getPostId() {
 		return postId;
 	}
 
-	public void setPostId(Integer postId) {
+	public void setPostId(Long postId) {
 		this.postId = postId;
 	}
